requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.0
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
billiard==4.2.1
boto==2.49.0
boto3==1.39.3
botocore==1.39.3
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.9.2
cron-descriptor==1.4.5
cryptography==45.0.5
cssutils==2.11.1

defusedxml==0.7.1
Django==5.1.6
django-auditlog==3.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-inlinecss==0.3.0
django-property-filter==1.3.0
django-rest-passwordreset==1.5.0
django-sendgrid-v5==1.3.0
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-access-policy==1.5.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
drf-writable-nested==0.7.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
execnet==2.1.1
factory_boy==3.3.3
Faker==37.4.0
frozenlist==1.7.0
future==1.0.0
httpie==3.2.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mirakuru==2.6.1
more-itertools==10.7.0
moto==5.1.8
msgpack==1.1.1
multidict==6.6.3
packaging==25.0
pendulum==3.1.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pynliner==0.8.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-factoryboy==2.8.1
pytest-faker==2.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.1
pytest-moto==0.2.0
pytest-sugar==1.0.0
pytest-xdist==3.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
responses==0.25.7
rich==14.0.0
rpds-py==0.26.0
s3transfer==0.13.0
sendgrid==6.12.4
service-identity==24.2.0
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
termcolor==3.1.0
tomli==2.2.1
twilio==9.6.4
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
xmltodict==0.14.2
yarl==1.20.1
zope.interface==7.2
pydantic[email]
